feat(Processing): Add columns to UMAP computations output + Various improvements
