+++ processing/processing/computations/compute_UMAPs.py
@@ -7,31 +7,58 @@ from processing.computations.iterate_timegroups import iterate_timegroups
 from processing.computations.load_features_for import load_features_for
 from processing.computations.timegroup_loaded_features import \
     timegroup_loaded_features
+from processing.utils.get_config import get_config
 
 
 def compute_UMAPs(cfg, plot, show):
     # we do the import here as it is very very slow
     import umap.umap_ as umap  # Why!!!!
     UMAP = umap.UMAP  # as we use "umap" a lot below
+
+    config = get_config()
+
     for umap_name, umap in cfg.umaps.items():
         print('... UMAP', umap_name, umap)
+
         for band in umap.bands:
             print('... ... BAND', band)
             dataset_times = []
             dataset_features = []
             dataset_labels = []
+            dataset_columns = []
+
             for r_name in umap.ranges:
                 r = cfg.ranges[r_name]
+
                 for s in umap.sites:
-                    range_times, range_features = load_features_for(cfg, band,
-                                                                    r, s)
+                    range_times, range_features = load_features_for(
+                        cfg,
+                        band,
+                        r,
+                        s
+                    )
+
+                    columns = [*config.files[s][3:]]
+
                     range_bins, group_starts = timegroup_loaded_features(
-                        range_times, r, umap.integration)
+                        range_times,
+                        r,
+                        umap.integration
+                    )
+
                     for g_start, g_end, t_start, g_start_i in iterate_timegroups(
-                            r, umap.integration, range_bins, group_starts):
+                            r,
+                            umap.integration,
+                            range_bins,
+                            group_starts,
+                    ):
                         dataset_times.append(t_start)
-                        dataset_features.append(
-                            np.mean(range_features[g_start:g_end, :], axis=0))
+                        dataset_columns.append(columns)
+
+                        dataset_features.append(np.mean(
+                            range_features[g_start:g_end, :], axis=0
+                        ))
+
                         dataset_labels.append(f'{r_name}/{s}')
 
             # print(np.shape(dataset_times), np.shape(dataset_features))
@@ -40,19 +67,22 @@ def compute_UMAPs(cfg, plot, show):
             if type(umap_random) is str:
                 umap_random = int(umap_random)
 
-            X = UMAP(
-                random_state=umap_random).fit_transform(
-                dataset_features)
+            X = UMAP(random_state=umap_random).fit_transform(dataset_features)
+
             out_path = pathlib.Path(cfg.variables['generated_base']).joinpath(
                 'umap', umap_name, band + '.json')
+
             out_path.absolute().parent.mkdir(parents=True, exist_ok=True)
+
             with open(out_path, "w") as jsonfile:
                 json.dump({
                     'X': X.tolist(),
                     't': [d.timestamp() for d in dataset_times],
                     'l': dataset_labels,
                     'binSize': umap.integration,
+                    'c': dataset_columns,
                 }, jsonfile)
+
             if plot:
                 import \
                     matplotlib.pyplot as plt  # import here to have matplotlib optional
