+++ processing/processing/globals.py
@@ -1,0 +1,1 @@
+all_sites = []
