--- processing/processing/utils/config/parse_config.py
@@ -1,151 +1,0 @@
-import collections
-import pathlib
-from datetime import datetime
-
-from pandas import pandas
-
-from processing.constants import EXCEL_COLUMNS
-from processing.utils.get_app_version import get_app_version
-from processing.utils.get_config_columns import get_config_columns
-from processing.utils.populate_empty_config_variables import \
-    populate_empty_config_variables
-
-all_sites = []
-
-
-def list_all_sites(files):
-    global all_sites
-
-    for f in files:
-        file = files[f]
-        site = file[0]
-
-        if site not in all_sites:
-            all_sites.append(site)
-
-
-def parse_sites(sites):
-    global all_sites
-    string = str(sites)
-
-    if string == 'nan':
-        return all_sites
-
-    return string.split(',')
-
-
-def digest_xtable_columns(xpath, xt, c_key, c_values=None, yield_type=None,
-                          disable_prefix=False, allow_duplicate=False):
-    if c_values is None:
-        c_values = ''
-    simple = False
-    if type(c_values) == str:
-        c_values = [c_values]
-        simple = True
-
-    if not disable_prefix:
-        c_values = [c_key + '_' + c for c in c_values]
-
-    types = []
-    for c_val in c_values:
-        if ':' in c_val:
-            t = c_val.split(':')[1]
-            types.append({
-                             'I': int,
-                             'D': lambda v: datetime.strptime(v, '%Y%m%d_%H%M'),
-                             'L': lambda v: v.split(','),
-                             'SITES': parse_sites,
-                             'L-': lambda v: v.split('-'),
-                             'L-D': lambda v: [
-                                 datetime.strptime(vv, '%Y%m%d_%H%M') for vv in
-                                 v.split('-')],
-                         }[t])
-        else:
-            types.append(str)
-
-    c_values = [c_val.split(':')[0] for c_val in c_values]
-
-    for c in [c_key] + c_values:
-        if c not in xt:
-            print(xt.columns)
-            raise Exception(f'In {xpath}, need a column named "{c}"')
-
-    res = {}  # actually not as we yield
-    for i in range(len(xt[c_key])):
-        if type(xt[c_key][i]) == str:
-            k = xt[c_key][i]
-            if k in res and not allow_duplicate:
-                raise Exception(
-                    f'In {xpath}: column "{c_key}" contains "{k}" twice.')
-            if simple:
-                v = [types[ic](xt[c][i]) for ic, c in enumerate(c_values)]
-                res[k] = v
-                yield k, v[0]
-            else:
-                start = 0 if disable_prefix else len(c_key) + 1
-                v = namedtuple_dic({c[start:]: types[ic](xt[c][i]) for ic, c in
-                                    enumerate(c_values)}, yield_type)
-                res[k] = v
-                yield k, v
-
-
-def namedtuple_dic(d, typename='new'):
-    top = collections.namedtuple(typename, [i for i, j in d.items() if
-                                            not i.startswith('_')])(
-        *[j for i, j in d.items() if not i.startswith('_')])
-    return top
-
-
-def parse_config(path='config.xlsx', sheet=0):
-    _xfile = pandas.ExcelFile(path)
-    _xtable = _xfile.parse(sheet, converters={'variables_': str})
-    _renaming = {i: i.split(' (')[0] for i in _xtable.columns if ' (' in i}
-    _xtable.rename(_renaming, inplace=True, axis='columns', errors="raise")
-
-    variables = dict(digest_xtable_columns(path, _xtable, 'variables'))
-
-    variables = populate_empty_config_variables(variables)
-
-    [unique_columns, _all_columns, _columns_length] = get_config_columns(
-        variables['audio_base'])
-
-    _files_selectors = ['site', 'start:D', 'tags:L']
-
-    for _i in range(_columns_length):
-        _files_selectors.append(EXCEL_COLUMNS[_i])
-
-    files = dict(digest_xtable_columns(
-        path,
-        _xtable,
-        'files',
-        _files_selectors,
-        'FILE'
-    ))
-
-    list_all_sites(files)
-
-    bands = dict(digest_xtable_columns(path, _xtable, 'bands'))
-
-    umaps = dict(digest_xtable_columns(
-        path,
-        _xtable,
-        'umaps',
-        ['integration:I', 'bands:L', 'sites:SITES', 'ranges:L'],
-        'UMAP'
-    ))
-
-    ranges = dict(digest_xtable_columns(path, _xtable, 'ranges', ':L-D'))
-
-    stringmap = dict(digest_xtable_columns(
-        path,
-        _xtable,
-        'stringmap',
-        ['to', 'color'],
-        'STRINGMAP'
-    ))
-
-    path = str(pathlib.Path(path).absolute())
-
-    app_version = get_app_version()
-
-    return namedtuple_dic(locals(), 'CFG')
