+++ processing/processing/utils/digest_config_columns.py
@@ -1,0 +1,2 @@
+def digest_config_columns(name, item):
+    return [item]
